feat: make `create-fuels` template app responsive
